Split shell streaming process runner
