+++ docs/DECISIONS.md
@@ -283,3 +283,4 @@
 - Sidebar thread list derivation lives in `QuillCodeSidebarThreadListBuilder.swift`, not in the aggregate `SidebarSurface` DTO. `SidebarSurface` keeps the stable public API consumed by native SwiftUI, static HTML, and search dialogs, while filtering, pinned/recent/archived partitioning, and date-bucket sectioning stay in the focused helper.
 - Native sidebar command payloads are built through `QuillCodeSidebarCommandAdapter.swift`. SwiftUI row and bulk-action views should not construct `WorkspaceCommandSurface` values inline; selection toggles and bulk actions share the same adapter so command IDs, titles, categories, and enabled state stay consistent.
 - Static HTML browser snapshots live in `BrowserHTMLSnapshotBuilder`; `BrowserInspector` remains the URL/file/fetched-page adapter and delegates title/count/outline/snippet extraction to the focused builder.
+- Streaming shell command lifecycle lives in `ShellStreamingProcessRunner`; `ShellToolExecutor` remains the public shell facade for blocking, cancellable, and streaming entry points while the runner owns async stdout/stderr readers, timeout termination, cancellation, and final streamed `ToolResult` emission.
